--- src/dataset/challenge/types/index.ts
@@ -1,5 +1,0 @@
-export interface DatasetChallenge {
-    leaves: string[]
-    siblings: string[][]
-    paths: string[]
-}
