--- src/dataset/requirement/types/index.ts
@@ -1,7 +1,0 @@
-export interface DatasetRequirement {
-    dataPreparers: string[]
-    storageProviders: string[]
-    regions: number
-    countrys: number
-    citys: number[]
-}
