+++ src/module/dataset/challenge/types/index.ts
@@ -1,0 +1,9 @@
+import { Entity } from "@unipackage/ddd"
+
+export interface DatasetChallenge {
+    leaves: string[]
+    siblings: string[][]
+    paths: string[]
+}
+
+export class DatasetChallenge extends Entity<DatasetChallenge> { }
