+++ test/shared/generator.ts
@@ -1,5 +1,5 @@
 import * as utils from "./utils"
-import { DatasetMetadata } from "../../src/dataset/metadata/types/index"
+import { DatasetMetadata } from "../../src/module/dataset/metadata/types/index"
 
 export class Generator {
